+++ hydrofunctions/__init__.py
@@ -31,23 +31,25 @@ List all of the different attributes and methods with dir()::
 
     >>> dir(response)
 
-Read more about Hydrofunctions here: http://hydrofunctions.readthedocs.io/
+Read more about Hydrofunctions here: https://hydrofunctions.readthedocs.io/
 
 """
 
 __title__ = 'hydrofunctions'
-__version__ = '0.1.4'
+__version__ = '0.1.5'
 __author__ = 'Martin Roberge'
 __email__ = 'mroberge.whois@gmail.com'
 __license__ = 'MIT'
-__copyright__ = 'Copyright 2016 Martin Roberge'
+__copyright__ = 'Copyright 2016 Martin Roberge and contributors'
 
 
 
 
 from .exceptions import HydroNoDataError
 from .hydrofunctions import (
-        get_nwis, extract_nwis_dict, extract_nwis_df
+        get_nwis, extract_nwis_df, nwis_custom_status_codes
         )
-from .station import Station
-from .typing import check_NWIS_station_id, check_datestr
+from .station import Station, NWIS
+from .typing import (
+        check_NWIS_site, check_datestr, check_NWIS_service, check_period
+        )