+++ tests/test_online_resources.py
@@ -19,11 +19,27 @@ from __future__ import absolute_import, print_function
 from hydrofunctions import hydrofunctions as hf
 
 
-def test_will_get_nwis_return_response():
+"""
+List of Valid URL's for NWIS
+paste these into a web browser or use curl to use these URL's.
+
+
+List of URL's that Return Errors or Specific Conditions
+
+400: didn't select a site
+ https://waterservices.usgs.gov/nwis/dv/?format=json%2C1.1&parameterCd=00060&startDT=2017-01-01&endDT=2017-12-31
+
+400: site names are too short to be right, should be 8 digits
+https://waterservices.usgs.gov/nwis/dv/?format=json%2C1.1&sites=51059%2C51061&parameterCd=00060&startDT=2017-01-01&endDT=2017-12-31
+
+400: wrote countyCd twice
+https://waterservices.usgs.gov/nwis/dv/?format=json%2C1.1&countyCd=51059&countyCd=51061&parameterCd=00060&startDT=2017-01-01&endDT=2017-12-31
+"""
+
+def will_NWIS_return_200_response():
     """This test should not be run as a part of the normal test suite.
     This is a resource test only!
-    Nose2 will find it.
-    Unittest will not.
+    Because of the function name, it is not discoverable by Unittest or pytest.
     """
 
     expected = 200
@@ -32,5 +48,10 @@ def test_will_get_nwis_return_response():
     assert expected == actual
     print('NWIS is up and running!')
 
+
+
+
+
+
 if __name__ == "__main__":
-    test_will_get_nwis_return_response()
+    will_NWIS_return_200_response()